move the create_app method to init file
